Pushed 20230602 Updatesgit add resources/views/members/show.blade.php

+++ database/migrations/2022_10_03_143706_create_accounts_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->year('year');
             $table->string('Name');
             $table->string('Code');
-            $table->float('AnualPrinciple');
+            $table->double('AnualPrinciple', 12, 2);
             $table->timestamps();
         });
     }
